sql: move TableColSet from schemaexpr package to catalog
The `TableColSet` type describes a set of columns of a table. This
commit moves the type to the `catalog` package because it its usage is
not specific to the functionality in the `schemaexpr` package.

Release note: None
