+++ pkg/sql/catalog/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "descriptor.go",
         "errors.go",
         "table_col_map.go",
+        "table_col_set.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/catalog",
     visibility = ["//visibility:public"],
@@ -34,6 +35,7 @@ go_test(
     srcs = [
         "dep_test.go",
         "descriptor_test.go",
+        "table_col_set_test.go",
     ],
     embed = [":catalog"],
     deps = [
@@ -42,6 +44,7 @@ go_test(
         "//pkg/sql/catalog/schemadesc",
         "//pkg/sql/catalog/tabledesc",
         "//pkg/testutils/buildutil",
+        "//pkg/util",
         "//pkg/util/leaktest",
         "//vendor/github.com/cockroachdb/redact",
         "//vendor/github.com/stretchr/testify/require",
